Add binary index support and fix scaling limits in Carta
- Add auto-detection of binary index files in tile server
- Support instant startup via mmap when loading pre-built index
- Change node/way map indices from uint32_t to size_t (supports >4B nodes)
- Fix thread-local render cache leak with pthread_key destructor
- Fix mongoose build on macOS (add _DARWIN_C_SOURCE)
- Document production workflow: build index from PBF, serve from index
- Document what's pre-computed in index and design tradeoffs

The binary index provides <1s startup vs 10-60s for PBF parsing,
while maintaining the same tile serving performance.

+++ carta/include/ct_types.h
@@ -466,7 +466,7 @@ typedef struct {
     /* Node ID to index lookup (hash map) */
     struct {
         int64_t *keys;
-        uint32_t *values;
+        size_t *values;     /* Index into nodes array (size_t for >4B nodes) */
         size_t capacity;
         size_t count;
     } node_map;
@@ -479,7 +479,7 @@ typedef struct {
     /* Way ID to index lookup (for relation member resolution) */
     struct {
         int64_t *keys;
-        uint32_t *values;
+        size_t *values;     /* Index into ways array (size_t for >4B ways) */
         size_t capacity;
         size_t count;
     } way_map;
